Fixed bug in jumps and routines lists and on emitter
List was being passed by value so the pointer allocation was lost. Fixed by passing the list as reference instead.

There was also a bug when creating a new list. The pointer was not initialized to NULL. Added create_functions which initialize the pointer correctly.

The logic emitting bytes or words arguments was wrong. Changed the parameter struct and used an enum to identify the type of opcode.

Reformatted opcodes file